Move commands to 'commands' package
